# openapi-appengine.yaml
swagger: "2.0"
info:
  description: "Ejemplo OpenAPI de Java Flexible dentro de Google Cloud Platform"
  version: "1.0.0"
  title: "GCP Demo"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "antonino.salas@cirrocode.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "cirrocode-dev.appspot.com"
basePath: "/gcp"
tags:
- name: "hola"
  description: "Hola Mundo!"
  externalDocs:
    description: "Encuentra mas en"
    url: "http://cirrocode.com"
schemes:
- "http"
paths:
  /hola/{nombre}:
    get:
      tags:
      - "hola"
      summary: "Saluda a una persona por su nombre"
      description: "Retorna un mensaje de saludo"
      operationId: "saludo"
      produces:
      - "application/json"
      parameters:
      - name: "nombre"
        in: "path"
        description: "Nombre de la persona a saludar"
        required: true
        type: "string"
      responses:
        200:
          description: "Operacion exitosa"
          schema:
            $ref: "#/definitions/HolaResponse"
      security:
      - api_key: []
  /mensajes:
    get:
      tags:
      - "mensajes"
      summary: "Recupera los mensajes de saludo"
      description: "Recupera mensajes"
      operationId: "leerMensajes"
      produces:
      - "application/json"
      responses:
        200:
          description: "Operacion exitosa"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/MensajesResponse"
      security:
      - api_key: []
  /mensajes/{id_saludo}:
    get:
      tags:
      - "mensajes"
      summary: "Busca un mensaje por su id"
      description: "Regresa un mensaje"
      operationId: "getMessageById"
      produces:
      - "application/json"
      parameters:
      - name: "id_saludo"
        in: "path"
        description: "Id del mensaje a regresar"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200: 
          description: "Operacion exitosa"
        400: 
          description: "Error al realizar la operacion"
        404: 
          description: "Mensaje no encontrado"
      security:
      - api_key: []
    put:
      tags:
      - "mensajes"
      summary: "Edita los mensajes de saludo"
      description: "Edita mensajes"
      operationId: "editMessageById"
      produces:
      - "application/json"
      parameters:
      - name: "id_saludo"
        in: "path"
        type: "integer"
        format: "int64"
        description: " "
        required: true
      - name: "Saludo"
        in: "body"
        description: "Nuevo saludo"
        required: true
        schema:
          type: "string"
      responses:
        201: 
          description: "Operacion exitosa"
        404: 
          description: "No encontrado"
        400: 
          description: "Error al realizar la operacion"
      security:
      - api_key: []
    delete:
      tags:
      - "mensajes"
      summary: "Elimina los mensajes de saludo"
      description: "Elimina mensajes"
      operationId: "deleteMessageById"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id_saludo"
        in: "path"
        type: "integer"
        format: "int64"
        description: " "
        required: true
      responses:
        200: 
          description: "Operacion exitosa"
        404: 
          description: "No encontrado"
        400: 
          description: "Error al realizar la operacion"
      security:
        - api_key: []
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "query"
definitions:
  HolaResponse:
    type: "object"
  MensajesResponse:
    type: "object"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"